+++ rastervision_pytorch_backend/rastervision/pytorch_backend/examples/semantic_segmentation/isprs_potsdam.py
@@ -17,6 +17,7 @@ def get_config(runner,
                processed_uri,
                root_uri,
                multiband=False,
+               augment=False,
                test=False):
     train_ids = [
         '2-10', '2-11', '3-10', '3-11', '4-10', '4-11', '4-12', '5-10', '5-11',
@@ -98,6 +99,75 @@ def get_config(runner,
     chip_options = SemanticSegmentationChipOptions(
         window_method=SemanticSegmentationWindowMethod.sliding, stride=chip_sz)
 
+    if augment:
+        mu = np.array((0.485, 0.456, 0.406, 0.485))[:len(channel_order)]
+        std = np.array((0.229, 0.224, 0.225, 0.229))[:len(channel_order)]
+
+        if multiband:
+            # not all transforms work with more than 3 channels, here are
+            # some of the ones that do
+            aug_transform = A.Compose([
+                A.Flip(),
+                A.Transpose(),
+                A.RandomRotate90(),
+                A.ShiftScaleRotate(),
+                A.FancyPCA(),
+                A.GaussNoise(),
+                A.OneOf([
+                    A.Blur(),
+                    A.MotionBlur(),
+                    A.Downscale(),
+                ]),
+                A.OneOf([
+                    A.GridDistortion(),
+                ]),
+                A.CoarseDropout(max_height=32, max_width=32, max_holes=5)
+            ])
+        else:
+            aug_transform = A.Compose([
+                A.Flip(),
+                A.Transpose(),
+                A.RandomRotate90(),
+                A.ShiftScaleRotate(),
+                A.OneOf([
+                    A.CLAHE(),
+                    A.FancyPCA(),
+                    A.HueSaturationValue(hue_shift_limit=10),
+                    A.RGBShift(),
+                    A.ToGray(),
+                    A.ToSepia(),
+                ]),
+                A.OneOf([
+                    A.RandomBrightness(),
+                    A.RandomGamma(),
+                ]),
+                A.OneOf([
+                    A.GaussNoise(),
+                    A.ISONoise(),
+                    A.RandomFog(),
+                ]),
+                A.OneOf([
+                    A.Blur(),
+                    A.MotionBlur(),
+                    A.ImageCompression(),
+                    A.Downscale(),
+                ]),
+                A.OneOf([
+                    A.GridDistortion(),
+                ]),
+                A.CoarseDropout(max_height=32, max_width=32, max_holes=5)
+            ])
+
+        base_transform = A.Normalize(mean=mu.tolist(), std=std.tolist())
+        plot_transform = A.Normalize(
+            mean=(-mu / std).tolist(),
+            std=(1 / std).tolist(),
+            max_pixel_value=1.)
+    else:
+        aug_transform = None
+        base_transform = None
+        plot_transform = None
+
     backend = PyTorchSemanticSegmentationConfig(
         model=SemanticSegmentationModelConfig(backbone=Backbone.resnet50),
         solver=SolverConfig(
@@ -109,7 +179,10 @@ def get_config(runner,
             one_cycle=True),
         log_tensorboard=True,
         run_tensorboard=False,
-        test_mode=test)
+        test_mode=test,
+        base_transform=A.to_dict(base_transform),
+        aug_transform=A.to_dict(aug_transform),
+        plot_options=PlotOptions(transform=A.to_dict(plot_transform)))
 
     if multiband:
         channel_display_groups = {'RGB': (0, 1, 2), 'IR': (3, )}