+++ rastervision_pytorch_backend/rastervision/pytorch_backend/pytorch_chip_classification_config.py
@@ -20,6 +20,9 @@ class PyTorchChipClassificationConfig(PyTorchLearnerBackendConfig):
         data.class_colors = pipeline.dataset.class_config.colors
         data.img_sz = pipeline.train_chip_sz
         data.augmentors = self.augmentors
+        data.base_transform = self.base_transform
+        data.aug_transform = self.aug_transform
+        data.plot_options = self.plot_options
 
         learner = ClassificationLearnerConfig(
             data=data,