+++ docs/changelog.rst
@@ -7,13 +7,14 @@ Raster Vision 0.13
 Features
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-* Added support for multiband images `#972 <https://github.com/azavea/raster-vision/pull/972>`_
+* Add support for multiband images `#972 <https://github.com/azavea/raster-vision/pull/972>`_
 * Add support for vector output to predict command `#980 <https://github.com/azavea/raster-vision/pull/980>`_
 * Add support for weighted loss for classification and semantic segmentation `#977 <https://github.com/azavea/raster-vision/pull/977>`_
 * Add multi raster source `#978 <https://github.com/azavea/raster-vision/pull/978>`_
 * Add support for fetching and saving external model definitions `#985 <https://github.com/azavea/raster-vision/pull/985>`_
 * Add support for external loss definitions `#992 <https://github.com/azavea/raster-vision/pull/992>`_
 * Upgrade to pyproj 2.6 `#1000 <https://github.com/azavea/raster-vision/pull/1000>`_
+* Add support for arbitrary albumentations transforms `#1001 <https://github.com/azavea/raster-vision/pull/1001>`_
 
 Bug Fixes
 ~~~~~~~~~~~~